+++ sdk/containerservice/arm-containerservice/src/operations/fleetMembers.ts
@@ -174,7 +174,8 @@ export class FleetMembersImpl implements FleetMembers {
     );
     const poller = new LroEngine(lro, {
       resumeFrom: options?.resumeFrom,
-      intervalInMs: options?.updateIntervalInMs
+      intervalInMs: options?.updateIntervalInMs,
+      lroResourceLocationConfig: "azure-async-operation"
     });
     await poller.poll();
     return poller;