+++ sdk/containerservice/arm-containerservice/src/operations/fleets.ts
@@ -207,7 +207,8 @@ export class FleetsImpl implements Fleets {
     );
     const poller = new LroEngine(lro, {
       resumeFrom: options?.resumeFrom,
-      intervalInMs: options?.updateIntervalInMs
+      intervalInMs: options?.updateIntervalInMs,
+      lroResourceLocationConfig: "azure-async-operation"
     });
     await poller.poll();
     return poller;